connected PostgreSQL and PgAdmin
I had installed PostgreSQL and PgAdmin before the changes.

Because I am not using the default MSSQL, I also updated the connection string in appsettings.json.

I Deleted the old migrations folder and made a new one using the console after adding the appropriate Nuget packages.

To verify that it worked, I used PgAdmin and verified that the database was created based on the connection string. The associated tables present and contain the custom columns.